Split V3Order into further part and decouple various components
