+++ src/Makefile_obj.in
@@ -266,6 +266,9 @@ RAW_OBJS_PCH_ASTNOMT = \
 	V3Name.o \
 	V3Order.o \
 	V3OrderGraphBuilder.o \
+	V3OrderParallel.o \
+	V3OrderProcessDomains.o \
+	V3OrderSerial.o \
 	V3Param.o \
 	V3Partition.o \
 	V3Premit.o \
